Switch tests to use fake data and run entirely locally
Also clean up some old cruft
