+++ .gitignore
@@ -1,3 +1,6 @@
+# Virtualenv
+venv
+
 # Build files
 *.pyc
 vulcanized.html
