--- data/test_api_keys.json
@@ -1,6 +1,0 @@
-{
-  "sendgrid" : {
-    "username": "sguser",
-    "password": "sgpass"
-  }
-}
