--- data/test_crypto.key
@@ -1,1 +1,0 @@
-baadf00dbaadf00dbaadf00dbaadf00d
