--- data/test_id_table.json
@@ -1,1 +1,0 @@
-[["tturtle", 9998], ["ccricketcreeper", 9997], ["sfassino", 9996], ["bbison", 9995], ["ggrasshopper", 9994], ["suzwack", 4093]]
