+++ requirements.txt
@@ -5,3 +5,4 @@ requests
 google-cloud-secret-manager
 google-cloud-storage
 Pillow
+pytest
